following the community-standard practices even further!
After a discussion with Pierre Courtieu, using a _CoqProject file
seems to be the future of Coq packaging.

Note the addition of Makefile.coq_makefile's "all" target to the
install step. It will re-compile the to-be-installed .v files in
a coq_makefile-controlled way, which adds the right "-R . Dblib"
parameter at compilation-time that will allow Dblib to be actually
usable as an external library (without them, Coq may complain about
disagreement between logical and physical path, or something like
this).